+++ attendispeechserviceexample/src/main/java/nl/attendi/attendispeechserviceexample/examples/screens/soapscreen/SoapScreenViewModel.kt
@@ -3,6 +3,7 @@ package nl.attendi.attendispeechserviceexample.examples.screens.soapscreen
 import android.content.Context
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -10,7 +11,6 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.runBlocking
 import nl.attendi.attendispeechservice.components.attendirecorder.plugins.AttendiAudioNotificationPlugin
 import nl.attendi.attendispeechservice.components.attendirecorder.plugins.AttendiErrorPlugin
 import nl.attendi.attendispeechservice.components.attendirecorder.plugins.AttendiStopOnAudioFocusLossPlugin
@@ -135,10 +135,19 @@ class SoapScreenViewModel(private val applicationContext: Context) : ViewModel()
     }
 
     override fun onCleared() {
-        // The reason runBlocking(Dispatchers.IO) is used here instead of CoroutineScope(Dispatchers.IO)
-        // is because onCleared() is a synchronous, blocking function, and Kotlin does not allow suspending
-        // functions or coroutine scopes directly in onCleared().
-        runBlocking(Dispatchers.IO) {
+        /**
+         * Releases the recorder resources owned by this ViewModel in a detached Coroutine scope.
+         *
+         * The ViewModel that creates recorder instances and their associated plugins is
+         * responsible for releasing them when they are no longer needed.
+         *
+         * This implementation performs the cleanup from [onCleared] because the example
+         * screen navigates away immediately. If needed, the process of releasing
+         * recorder resources can be achieved on a custom scope using the suspend function [recorder.release].
+         * Check [TwoMicrophonesStreamingScreenViewModel] how the recorder release is handled
+         * and embed into the screen lifecycle.
+         */
+        CoroutineScope(Dispatchers.IO).launch {
             recorder.release()
         }
         super.onCleared()